move tooltip icon from asset to tour directory
